cleaning types
