+++ src/Core/Domain/Contracts/Task/TextGeneration/Models.cs
@@ -7,10 +7,3 @@ public class GenerationResponse
     [JsonPropertyName("generated_text")]
     public string GeneratedText { get; set; } = string.Empty;
 }
-
-public sealed class TextGenerationResponseData
-{
-    public string RequestId { get; set; } = string.Empty;
-    public string Error { get; set; } = string.Empty;
-    public GenerationResponse Response { get; set; }
-}
