+++ src/Core/Domain/Contracts/Task/Translation/Models.cs
@@ -13,10 +13,3 @@ public class TranslationResponse
     [JsonPropertyName("target_language")]
     public string TargetLanguage { get; set; } = string.Empty;
 }
-
-public sealed class TranslationResponseData
-{
-    public string RequestId { get; set; } = string.Empty;
-    public string Error { get; set; } = string.Empty;
-    public TranslationResponse Response { get; set; }
-}
