+++ src/Core/Presentation/DependencyInjection.cs
@@ -24,7 +24,6 @@ public static class DependencyInjection
     public static IEndpointRouteBuilder AddEndPoints(this IEndpointRouteBuilder app)
     {
         app.AddHealthCheckEndPoints();
-        app.AddModelsEndPoints();
         app.AddTasksEndPoints();
 
         return app;
